backend/RoadBuddy/requirements.txt
APScheduler==3.11.0
bidict==0.23.1
blinker==1.9.0
CacheControl==0.14.2
cachetools==5.5.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.0
dnspython==2.7.0
eventlet==0.39.0
exceptiongroup==1.2.2
firebase-admin==6.6.0
Flask==3.1.0
Flask-Cors==5.0.0
Flask-SocketIO==5.5.1
google-api-core==2.24.0
google-api-python-client==2.159.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.20.0
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.69.0
grpcio-status==1.69.0
gunicorn==23.0.0
h11==0.14.0
httplib2==0.22.0
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
msgpack==1.1.0
packaging==24.2
pluggy==1.5.0
proto-plus==1.25.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.1
pyparsing==3.2.1
pytest==8.3.4
python-dotenv==1.0.1
python-engineio==4.11.2
python-socketio==5.12.1
pytz==2025.1
requests==2.32.3
requests-mock==1.12.1
rsa==4.9
simple-websocket==1.1.0
stripe==11.5.0
tomli==2.2.1
typing_extensions==4.12.2
tzlocal==5.3
uritemplate==4.1.1
urllib3==2.3.0
Werkzeug==3.1.3
wsproto==1.2.0
